feat: recommendation engine expansion, ERC-4337 bundler, post-quantum crypto, and architecture cleanup
Major uncommitted WIP consolidated into one commit: candidate repository,
coalescer, scoring/weights modules, feedsource impls, Nebula DLQ/recorder/
transport layers, Dilithium3 crypto + Rustler NIF bindings, bundler
reputation throttling, thera_friendz indexer, event_processor module split,
and 8 new migrations.

Also folds in this session's architecture-review fixes: duplicate indexer
removal, shared schema_consts helpers (vid/rank/vertex-upsert dedup fixing
a live comments_on rank divergence bug between REST and Kafka write paths),
write_edge/delete_edge consolidation in graph_client.rs (verified via new
golden-string tests), and removal of 4552 stale target_docker/ build
artifacts that were accidentally tracked (now gitignored).

+++ src/bundler/paymaster.rs
@@ -54,8 +54,8 @@ impl PaymasterSigner {
     pub fn sign_paymaster_data(
         &self,
         user_op: &PackedUserOperation,
-        valid_until: u32,
-        valid_after: u32,
+        valid_until: u64,
+        valid_after: u64,
     ) -> Result<Bytes> {
         let hash: B256 = compute_paymaster_hash(
             user_op,
@@ -73,10 +73,10 @@ impl PaymasterSigner {
         out.extend_from_slice(&self.pm_verification_gas_limit.to_be_bytes());
         out.extend_from_slice(&self.pm_post_op_gas_limit.to_be_bytes());
 
-        let vu = (valid_until as u64).to_be_bytes();
+        let vu = valid_until.to_be_bytes();
         out.extend_from_slice(&vu[2..]);
 
-        let va = (valid_after as u64).to_be_bytes();
+        let va = valid_after.to_be_bytes();
         out.extend_from_slice(&va[2..]);
 
         out.extend_from_slice(&sig_bytes);
